abort if there's a local .git directory
If ga is called accidently from within a local .git dir, then it
create a new cachefile that then gets lingers in the sandbox.
Usually this is a mistake. Therefor ga aborts in this situation.
Create the cachefile manually to work around this condition.
In future a --force switch is planned, as well as a cache update
feature and the possibility to exclude files